codacy improvement to branch
